requirements.txt
python-dotenv
langchain-community 
langchain-core
openai
langchain-huggingface